[RUNTIME][BUGFIX] Fix DSO module problem when its parent get destructed. (#7918)
* [RUNTIME][BUGFIX] Fix DSO module problem when its parent get destructed.

Previouslyw we set the context of the DSO module to be the root module.
This can cause problem when the root module is not the dso module and
get destructued early (but we still need the dso module).

This PR makes the following change to fix the problem.

- Merge multiple DSO modules to one during export.
- Set the context to be the (only one) dso module.
- Updated testcase to cover the problem.

The enhancement creates some restrictions on the dso import hierachy
(all dso modules needs to be merged without a cycle). This is the case
for our current use scenario. The merged logic is also more consistent
as the library itself is merged.

* Address review comments.