[Metal] Fix bad stream after interrupted tuning session (#8244)
* [Metal] Fix bad stream after interrupted tuning session

After interrupted tuning session, we may face the problem that the
stream object was released, but we didn't create a new one. In this case
it wasn't possible to run a new Metal task on the device without
restarting rpc application.

Created a global function `metal.ResetGlobalState` which should be
called in RPC application when the connection was closed. In this
function, we reinitialize the streams of Metal devices. And it
guarantees to us that the new RPC session will work with the correct
streams.

* Refactor metal_device_api

- Rename function GetStream -> CastStreamOrGetCurrent
- Add several checks on device id
- When we use `SetStream` with nullptr, then the default stream will be
  associated with the device.