[Vulkan] Remote target.h #include (#8813)
Was added in #8127, should have been removed in #8171 along with the
rest of the references outside of libtvm_runtime.so.  This didn't
impact the Vulkan+g++ builds, because no symbols were accessed outside
of the runtime library.  However, it broke the Vulkan+Windows builds,
which expected symbols due to the `__declspec(dllexport)` defintion of
`TVM_DLL` on MSVC (see #8805).  This wasn't caught by the CI build on
Windows, because it doesn't perform the Vulkan build.