[TVMC] run: Don't use static path to find model.tar (#9712)
* [TVMC] run: Don't use static path to find model.tar

Currently 'tvmc run' when '--device micro' is specified looks for the
model in the project directory at <project_dir>/model.tar. That works
for Zephyr but fails on Arduino because model.tar is actually located at
<project_dir>/src/model/model.tar. As a consequence 'tvmc run' when used
to run a model on Arduino exists because model.tar is never found.

This commit fixes it by using the MLF path returned by the Project API
instead of using a static path.

This commit also adds a project_dir attribute to TVMCPackage that can be
set when a MLF archive is loaded/imported so the project dir can be
conveniently found (similarly to package_path attribute).

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>

* [TVMC] test: Add test for importing a MLF with project_dir

Add test for TVMCPackage when importing a MLF archive and setting a
project directory too. Setting a project dir is only supported when a
MLF model is imported, so it must fail on Classic format.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>