[microNPU][2a] Add CascaderGraph for cascading analysis (#9469)
A CascaderGraph augments a TE graph with additional
information needed by the cascading algorithms. This
includes defining a strict ordering on the operators
as well as including all the Propagators needed to
do the affine analysis of cascades.

The CascaderGraph consists of two object types, Parts
and Tensors. A Part is an augmented operator which
includes the Propagators and a Tensor is similar to a
TE tensor but stores additional information like
compression ratio.