Add Action to add cc'ed people as reviewers (#9934)
* Add action to label mergeable PRs

Developers often have to ping a committer once their PRs are both passing in CI and are approved. This helps facilitate this process by marking such PRs with a label `ready-for-merge` so committers can easily filter for outstanding PRs that need attention.

* Fix lint and add tests

* Add Action to add cc'ed people as reviewers

This provides a mechanism for non-triager/reviewer/committer PR authors to request reviews through GitHub. Anyone that is referenced by `cc @username` in a PR body will be added as a reviewer (GitHub will limit the reviewers to those with actual permissions to leave reviews so the script to add can be simple).

* remove merge bot stuff

* Fix target triggers

Co-authored-by: driazati <driazati@users.noreply.github.com>