[BugFix] Fix NeedsMultiLevelTiling by skipping trivial block iterators (#10804)
This PR fixes a bug of `NeedsMultiLevelTiling`, which didn't consider the effect of trivial block iterators (iterators whose domains are `[0, 1)`). Such iterators impacts the following analysis by overlargely counting the number of iterators that are not used to index the block read regions, and might lead to the application of multi-level tiling where the rule is supposed not to apply.

To fix the problem, we simply skip such trivial block iterators.