[build] Update libinfo and add lint rule (#10774)
* [build] Update libinfo and add lint rule

This updates `tvm.support.libinfo()` to be in-line with the current tvm options. It also adds a lint rule to ensure these stay matched up in the future as well as a script to print out the options in more detail. This should add in communication when debugging (i.e. tell someone to run `python -c 'import tvm; tvm.support.describe()` to learn everything you need about their envrionment)

* Fix pylint

Co-authored-by: driazati <driazati@users.noreply.github.com>