[AOT] Enable A-Normal Form in the AOT executor (#11091)
* [AOT] Enable A-Normal Form in the AOT executor

The sequence of calls produced by the AOT executor codegen is arbitrary,
especially in the presence of 'branchy' networks. This makes it
difficult to analyze memory usage for each call. By running the
ToANormalForm pass to insert a series of let bindings before the
lowering and codegen stages, we can establish an ordering for the
evaluation of the external calls, thus allowing reliable analysis of
memory usage.

Change-Id: Ic320b68cde83c96b228a8d1d2829a0e8ac7b768f

* Maintain GetStorage(var) == GetStorage(value) invariant for lets

Change-Id: Id40b70f67a3e37f75b8331aa89f1819072e4d48e

* Add check to ensure ANF runs in AOT

Change-Id: I8de2bd19c7c17057e2bc89f6a68595780c2e9433

* Avoid let block traversal and don't visit var in let visitation

Change-Id: I74c080e2a09e84a75400db5c3395d508697d5d0f