[TIR][Arith] Additional Simplifications Inside Conditionals (#11524)
* [TIR][Arith] Use equality constraints in analyzer

Previously, constraints with inequalities were recognized and used for
simplifications by `ConstIntBoundAnalyzer` and `ModularSetAnalyzer`,
but constraints with equalities were not.  This adds equality-based
constraints.  (e.g. Inside the then-case of `if i==5`, the value of
`i` is known to be 5.)

* [TIR][Arith] RewriteSimplifier, apply literal constraints

Previously, constraints were only checked within a `tir.likely`
annotation.  After this change, constraints are used for
simplification of all boolean expressions.  (e.g. Within a conditional
`if i==n`, the expression `(i==n) and (j==m)` can be simplified to
`j==m`.)

* [TIR][Arith] Do not apply literal constraints to BufferLoad

If a literal constraint relies on the contents of a buffer, the
constraint may not be assumed to hold.  This prevents the incorrect
rewriting of `A[i]==n` to true within a `if A[i]==n` conditional, as
the value of `A[i]` may have changed.

* [TIR][Arith] Use each independent constraints in RewriteSimplifier

Inside a constraint `if i==n and j==m`, both `i==n` and `j==m` may be
replaced with true, even in separate expressions.

This commit uses a new internal utility function
`tvm::arith::ExtractConstraints`, which breaks up a boolean expression
into a list of true statements.  This may be used to reduce
duplication elsewhere, such as `const_int_bound.cc` and
`iter_affine_map.cc`.

* [TIR][Arith] Check for negation of literal constraints

When inside a conditional of `i!=n`, in addition to the previous
replacement of `i!=n` with true, we can also replace `i==n` with
false.

* [TIR][Arith] Added unittests for new simplifications

* Fix lint error

* Fixed handling of negation of non-boolean types

* Removed extra asterisk