[AOT] Calculate used memory at the callsite of primitive functions (#11208)
* [AOT] Calculate used memory at the callsite of primitive functions

Introduces a new pass in the AOT executor called "AnnotateUsedMemory"
which applies liveness analysis to the callsite of each primitive
function in order to calculate the total size of the live tensors at
this point of execution. The result is provided as a function annotation
called "used_memory", which can be consumed by later stages of the
compiler (e.g. external codegens) to provide more information about the
current memory consumption. This can be useful for some optimizations.

Change-Id: I8d6b7447498f19260358bbefe34029ddd86b9c89

* small fix to file description

Change-Id: I0e460f6cf43f9b12ffa5fc66fcb68e55304daeb2

* Various improvements addressing comments

In addition, a new "io_used_memory" annotation is added to the main
function which refers to the total size of the IO tensors in the
provided module, enabling these to be discounted from memory pressure
calculations where necessary.

Change-Id: Iafe9c85d7fc69c77a2115ed4efe7645160387c86

* addressing comments

Change-Id: I00f5ba80d5e004076e4c27d39bec143178b3b1dd

* add note for dynamic shapes

Change-Id: If6409e2953addfc880bcc6d95083b78bdf5a23d0