[microTVM][Zephyr] Fix missing BOARD in CMakeCache file (#12338)
This bug was caught in the microTVM hardware in the loop CI, because
only in the HW in the loop CI it's possible to test 'open_transport' and
ZephyrSerialTransport() method, i.e. run a model via the serial
transporter.                                  

This commit changes the microTVM Zephyr template project to read BOARD
from cmake file instead of cmake cache file.