[TIR][StorageRewrite] Allow in-place buffer reuse of non-flat memory (#12655)
* [TIR][StorageRewrite] Allow in-place buffer reuse of non-flat memory

Previously, shared buffer use was entirely disabled for non-flat
memory, since the existing checks for shared memory assume flat 1-d
spaces.  This was enforced in `FindAlloc` and validated in
`PrepareNewAlloc`.  The validation in `PrepareNewAlloc` could trigger,
if the buffer sharing was due to an in-place operation, and not
through the `FindAlloc` function.

In-place operations do not require N-d packing, nor do they introduce
ambiguity in how different code generators may interpret non-flat
physical indices.  Therefore, this commit relaxes the validation in
`PrepareNewAlloc`, allowing buffer reuse of non-flat buffers for
in-place operations.

* Update new StorageRewrite with correct allocate/buffer_decl usage