[ETHOSN] Transpose fully connected weights (#12970)
* [ETHOSN] Transpose fully connected weights

The NPU driver stack expects weights in IO (HWIO) format, however, Relay
uses an OI representation. Although the shape of the weight tensor was
correctly changed during codegen, the values in the weights tensor were
not being transposed. This lead to an output mismatch when the output
"units" was > 1. The tests didn't catch this due to using a weights
tensor of all 1's.

Change-Id: I51b2bcd14b677280ef3b6a6845d56b7dfacc7d6a

* Address comments

* Refactor use of weight transpose to common file between contrib
codegens.
* Make function areguments more explicit.
* Update network hashes.

Change-Id: Ib53bc7d2837b62908b92fd09062cbe9a8bb4ab30

* Fix lint

Change-Id: I6a1d9ffa8e3a747b7c77c9b27aa1b1c0d4c5cbff

* Fix cmsis-nn weights transpose

Change-Id: Ie89e429da222ffe17bc8faf831bf59217008a68a

* Address comments

Change-Id: Ie5ded2db3024b9e2c5095f01adea65798fc1da55