[LOGGING] Optionally print backtrace on segfault (#12959)
Add BACKTRACE_ON_SEGFAULT cmake option to install a signal handler to print a backtrace on segfault. Disabled by default because existing signal handlers may be overriden by this signal handler. Creating a backtrace may allocate, which should not happen in a signal handler, but the program is crashing so it shouldn't make any difference.

Hopefully this will help diagnose segfaults like in #12955.