[Hexagon] [runtime] Allow creation of thread manager without hardware resources (#13174)
This allows the creation of a thread manager with optional hardware resources.  This allows us to create a new object for unit tests without hardware resources.

This will fix the test instability we've seen recently, as all tests were using the global thread manager and things could be in an unknown state.