[TVMScript] Use new variable frame in If/Then/Else (#14250)
Previously, while TVMScript introduces a new scope for other
contexts (e.g. `for`, `while`, `with`, etc), the `if` and `else`
blocks did not introduce a new scope.  This caused erroneous parsing
errors if the `if` and `else` blocks each contained a variable with
the same name.  Added a `self.var_table.with_frame()` context resolves
this issue.