[TIR] [Schedule] Add get_output_blocks primitive (#14490)
* [TIR] [Schedule] Add get_output_blocks primitive

When scheduling fused ops, its really useful to be able to get all the
output blocks and schedule other blocks in the fused op with respect to
output blocks. This helps avoid hardcoding block names in manually
written schedules in many cases

* Fix cpplint errors

* Another lint

* Update primitive to take scope block as argument

This commit updates the primitive to take scope block as argument and
return all output blocks under that scope. In order to get all output
blocks of a PrimFunc, the root block can be passed