[TIR] Use String instead of StringImm for AttrStmtNode::node (#14491)
`tir::StringImm` can round-trip through TVMScript when used in a
context that requires a PrimExpr, such as the arguments of a
`tir::Call`.  However, contexts that only require a `ObjectRef`, such
as the `AttrStmtNode::node`, use the same TVMScript representation as
`"string_value"`, but are parsed `tvm::String` instances.

This commit updates `MakePackedAPI` to use `String` instead of
`StringImm` in its default value for `AttrStmtNode::node`.