[Node] Allow alternative root names in ObjectPath::Root() (#14569)
* [Node] Allow alternative root names in ObjectPath::Root()

Previously, the `ObjectPath` utility allowed tracking of an object's
location within a tree-like structure.  However, the base of the path
structure was hard-coded to be the string `<root>`.  For use cases
such as `StructuralEqual`, there is a clear root node.  However, other
cases such as using `ObjectPath` to specify an object's location
relative to a known variable, would require using that known
variable's name as the root, rather than the hard-coded string
`<root>`.

This commit adds an optional parameter to provide an alternative name
for the root node, to allow for these use cases.

* Updated python API, added unit test