[TIR] Restrict tir.transform.CombineContextCall to host functions (#14945)
* [Target] Added utility method TargetNode::HasKey()

This utility method makes it easier to determine if a target contains
a specific key.

* [TIR] Added utility method tvm::tir::IsHostFunc(const PrimFunc&)

For modules that contain both host and device functions, this utility
function checks whether a given PrimFunc is a host function, based on
the target annotation.

* [TIR] Restrict tir.transform.CombineContextCall to host functions

Previously, the `tir.transform.CombineContextCall` pass applied to all
functions in an `IRModule`, but was only applied to modules that
contain only host functions.  This commit updates
`tir.transform.CombineContextCall` to apply only to host functions.