[TIR] Restrict tir.transform.LowerTVMBuiltin to host functions (#14944)
* [Bugfix][TIR][VTA] Update host-side target, even without device func

This resolves an issue introduced by the combination of
https://github.com/apache/tvm/pull/14918 and
https://github.com/apache/tvm/pull/14945.  The bug occurred for
targets that do not require device-side codegen, but do require a
`device_type` other than `kDLCPU`.  It wasn't caught by CI, as the
issue only occurred with the combination of both PRs.

1. #14918 updated `SplitHostDevice` to only modify the `"target"`
   attribute when a device-side function has been extracted.

2. For VTA, there is no device-side function, as everything is done
   through host-side API calls.

3. From (1) and (2), the VTA examples kept the target
   `T.target("ext_dev", host="llvm")` after the `SplitHostDevice`
   pass, instead of being updated to `T.target("llvm")`.

4. #14945 restricted CombineContextCall to only apply to host-side
   passes.

5. From (4) and (5), the `CombineContextCall` pass was no longer
   applied to the VTA context calls.

This PR fixes `SplitHostDevice`, updating the target from
`T.target("ext_dev", host="llvm")` to `T.target("llvm")`, even if no
device sections have been extracted from the function.

* [TIR] Restrict tir.transform.LowerTVMBuiltin to host functions

Previously, the `tir.transform.LowerTVMBuiltin` pass applied to all
functions in an `IRModule`, but was only applied to modules that
contain only host functions.  This commit updates
`tir.transform.LowerTVMBuiltin` to apply only to host functions.

* Updated "stackvm" target to have "cpu" key.

With the presence/absence of the "cpu" key in a target used to
determine whether host-only calls should be run, should make sure to
add it to "stackvm".

* Update IsHostFunc() to use "host" tag instead of "cpu"

Current CI failures due to LowerTVMBuiltin not running on "hexagon"
target, and would like to avoid conflating cpu/host.

* Avoid "host" tag for now

* Update HEXAGON_AOT_LLVM_TARGET to be recognized as host