[TIR] Handle callees on same target, different codegen (#14988)
* [TIR] Handle callees on same target, different codegen

Prior to this commit, any caller that uses a different `Target` than
its callee is lowered to a device-kernel launch.  However, if the
caller and callee are on the same device, despite using a different
target (e.g. `Target("llvm")` and `Target("c")` both use `kDLCPU`),
then the kernel launch is unnecessary.

This commit updates `LowerDeviceKernelLaunch` to produce a kernel
launch only when the callee is on another device, and to produce
`T.call_extern` for callees on the same device.

* Rename "extern_method_call_" to "extern_function_call_"